requirements.txt
fastapi
uvicorn
openai
python-telegram-bot
python-dotenv